split reading in another file
