+++ src/lib.rs
@@ -1,5 +1,7 @@
 pub mod address;
 pub mod formatter;
+pub mod read_configuration;
 
 pub use address::Address;
 pub use formatter::Formatter;
+
